+++ tests/unit/customizations/s3/test_ls_command.py
@@ -14,6 +14,7 @@
 from awscli.testutils import BaseAWSCommandParamsTest
 from dateutil import parser, tz
 
+
 class TestLSCommand(BaseAWSCommandParamsTest):
 
     def test_operations_used_in_recursive_list(self):
@@ -34,6 +35,7 @@ class TestLSCommand(BaseAWSCommandParamsTest):
         self.assertEqual(
             stdout, '%s        100 foo/bar.txt\n'%time_local.strftime('%Y-%m-%d %H:%M:%S'))
 
-
-if __name__ == "__main__":
-    unittest.main()
+    def test_errors_out_with_extra_arguments(self):
+        stderr = self.run_cmd('s3 ls --extra-argument-foo', expected_rc=255)[1]
+        self.assertIn('Unknown options', stderr)
+        self.assertIn('--extra-argument-foo', stderr)